Address and enforce compiler warnings
- Increase the compiler level from 3 to 4
- Enable treating warnings as errors
- Remove some extant warning-disabling pragmas
- Fix outstanding warnings

+++ src/wxProfilerGUI/CallstackView.h
@@ -64,7 +64,7 @@ class CallstackView : public wxWindow
 	size_t									callstackActive;
 	wxString								callstackStats;
 	const Database::Symbol					*currSymbol;
-	long									itemSelected;
+	size_t									itemSelected;
 
 	void setupColumn(ColumnType id, int width, const wxString &name);
 	void updateList();
